Add new parameterized init task
Add a new gulp task 'plugin:initParameterized' to allow for the configuration json file to be created by passing in command line agruments instead of requiring a prompt.  

This will be useful, as it will make it much easier to run the init task in bash or powershell scripts.  Instead of having to prompt the user for input each time the init task is ran, the appServerPath and url can now be stored in variables and passed to the parameterized init.  

I would be willing to write unit tests if required to be merged, however, I'm not sure what the best way to mock receiving command line arguments would be, since I'm currently reading directly from yargs.